BFT-476: Refactor consensus tasks and storage
