+++ core/node/consensus/src/testonly.rs
@@ -49,7 +49,8 @@ use zksync_web3_decl::client::{Client, DynClient, L2};
 
 use crate::{
     batch::{L1BatchCommit, L1BatchWithWitness, LastBlockCommit},
-    en, ConnectionPool,
+    en,
+    storage::ConnectionPool,
 };
 
 /// Fake StateKeeper for tests.
