+++ core/node/node_framework/src/implementations/layers/consensus.rs
@@ -37,7 +37,7 @@ pub enum Mode {
 /// ## Adds tasks
 ///
 /// - `MainNodeConsensusTask` (if `Mode::Main`)
-/// - `FetcherTask` (if `Mode::External`)
+/// - `ExternalNodeTask` (if `Mode::External`)
 #[derive(Debug)]
 pub struct ConsensusLayer {
     pub mode: Mode,
@@ -101,7 +101,7 @@ impl WiringLayer for ConsensusLayer {
                     }
                 };
 
-                let task = FetcherTask {
+                let task = ExternalNodeTask {
                     config,
                     pool,
                     main_node_client,
@@ -130,7 +130,7 @@ impl Task for MainNodeConsensusTask {
 
     async fn run(self: Box<Self>, mut stop_receiver: StopReceiver) -> anyhow::Result<()> {
         // We instantiate the root context here, since the consensus task is the only user of the
-        // structured concurrency framework (`MainNodeConsensusTask` and `FetcherTask` are considered mutually
+        // structured concurrency framework (`MainNodeConsensusTask` and `ExternalNodeTask` are considered mutually
         // exclusive).
         // Note, however, that awaiting for the `stop_receiver` is related to the root context behavior,
         // not the consensus task itself. There may have been any number of tasks running in the root context,
@@ -151,7 +151,7 @@ impl Task for MainNodeConsensusTask {
 }
 
 #[derive(Debug)]
-pub struct FetcherTask {
+pub struct ExternalNodeTask {
     config: Option<(ConsensusConfig, ConsensusSecrets)>,
     pool: ConnectionPool<Core>,
     main_node_client: Box<DynClient<L2>>,
@@ -160,21 +160,21 @@ pub struct FetcherTask {
 }
 
 #[async_trait::async_trait]
-impl Task for FetcherTask {
+impl Task for ExternalNodeTask {
     fn id(&self) -> TaskId {
         "consensus_fetcher".into()
     }
 
     async fn run(self: Box<Self>, mut stop_receiver: StopReceiver) -> anyhow::Result<()> {
         // We instantiate the root context here, since the consensus task is the only user of the
-        // structured concurrency framework (`MainNodeConsensusTask` and `FetcherTask` are considered mutually
+        // structured concurrency framework (`MainNodeConsensusTask` and `ExternalNodeTask` are considered mutually
         // exclusive).
         // Note, however, that awaiting for the `stop_receiver` is related to the root context behavior,
         // not the consensus task itself. There may have been any number of tasks running in the root context,
         // but we only need to wait for stop signal once, and it will be propagated to all child contexts.
         let root_ctx = ctx::root();
         scope::run!(&root_ctx, |ctx, s| async {
-            s.spawn_bg(consensus::era::run_en(
+            s.spawn_bg(consensus::era::run_external_node(
                 ctx,
                 self.config,
                 self.pool,
